moved more cs constants
