+++ fast64_internal/oot/oot_constants.py
@@ -629,58 +629,6 @@ ootEnumSceneSetupPreset = [
     ("All Non-Cutscene Scene Setups", "All Non-Cutscene Scene Setups", "All Non-Cutscene Scene Setups"),
 ]
 
-ootEnumCSWriteType = [
-    ("Custom", "Custom", "Provide the name of a cutscene header variable"),
-    ("Embedded", "Embedded", "Cutscene data is within scene header (deprecated)"),
-    ("Object", "Object", "Reference to Blender object representing cutscene"),
-]
-
-ootEnumCSListType = [
-    ("Textbox", "Textbox", "Textbox"),
-    ("FX", "Scene Trans FX", "Scene Trans FX"),
-    ("Lighting", "Lighting", "Lighting"),
-    ("Time", "Time", "Time"),
-    ("PlayBGM", "Play BGM", "Play BGM"),
-    ("StopBGM", "Stop BGM", "Stop BGM"),
-    ("FadeBGM", "Fade BGM", "Fade BGM"),
-    ("Misc", "Misc", "Misc"),
-    ("0x09", "Cmd 09", "Cmd 09"),
-    ("Unk", "Unknown Data", "Unknown Data"),
-]
-
-ootEnumCSListTypeIcons = [
-    "ALIGN_BOTTOM",
-    "COLORSET_10_VEC",
-    "LIGHT_SUN",
-    "TIME",
-    "PLAY",
-    "SNAP_FACE",
-    "IPO_EASE_IN_OUT",
-    "OPTIONS",
-    "EVENT_F9",
-    "QUESTION",
-]
-
-ootEnumCSTextboxType = [("Text", "Text", "Text"), ("None", "None", "None"), ("LearnSong", "Learn Song", "Learn Song")]
-
-ootEnumCSTextboxTypeIcons = ["FILE_TEXT", "HIDE_ON", "FILE_SOUND"]
-
-ootEnumCSTransitionType = [
-    ("1", "To White +", "Also plays whiteout sound for certain scenes/entrances"),
-    ("2", "To Blue", "To Blue"),
-    ("3", "From Red", "From Red"),
-    ("4", "From Green", "From Green"),
-    ("5", "From White", "From White"),
-    ("6", "From Blue", "From Blue"),
-    ("7", "To Red", "To Red"),
-    ("8", "To Green", "To Green"),
-    ("9", "Set Unk", "gSaveContext.unk_1410 = 1, works with scene xn 11/17"),
-    ("10", "From Black", "From Black"),
-    ("11", "To Black", "To Black"),
-    ("12", "To Dim Unk", "Fade gSaveContext.unk_1410 255>100, works with scene xn 11/17"),
-    ("13", "From Dim", "Alpha 100>255"),
-]
-
 ootEnumDrawConfig = [
     ("Custom", "Custom", "Custom"),
     ("SDC_DEFAULT", "Default", "Default"),
